Split operations into their own file, akin to searchengine
